+++ R/confint-methods.R
@@ -34,9 +34,9 @@
 ##' load_mgcv()
 ##' \dontshow{
 ##' set.seed(2)
-##' op <- options(digits = 5)
+##' op <- options(digits = 3, cli.unicode = FALSE)
 ##' }
-##' dat <- gamSim(1, n = 400, dist = "normal", scale = 2)
+##' dat <- gamSim(1, n = 1000, dist = "normal", scale = 2)
 ##' mod <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")
 ##'
 ##' ## first derivatives of all smooths...
@@ -47,8 +47,10 @@
 ##' head(ci)
 ##'
 ##' ## simultaneous interval for smooth term of x1
+##' \dontshow{
 ##' set.seed(42)
-##' x1.sint <- confint(fd, parm = "x1", type = "simultaneous", nsim = 1000)
+##' }
+##' x1.sint <- confint(fd, parm = "x1", type = "simultaneous", nsim = 2500)
 ##' head(x1.sint)
 ##' \dontshow{options(op)}
 `confint.fderiv` <- function(object, parm, level = 0.95,
@@ -166,7 +168,7 @@
 ##'   vector of terms. If missing, all parameters are considered, although this
 ##'   is not currently implemented.
 ##' @param level numeric, `0 < level < 1`; the confidence level of the point-wise
-##'   or simultaneous interval. The default is `0.95` for a 95\% interval.
+##'   or simultaneous interval. The default is `0.95` for a 95% interval.
 ##' @param newdata data frame; containing new values of the covariates used in
 ##'   the model fit. The selected smooth(s) wil be evaluated at the supplied
 ##'   values.
@@ -210,19 +212,27 @@
 ##' @export
 ##'
 ##' @examples
-##'load_mgcv()
+##' load_mgcv()
+##' \dontshow{
 ##' set.seed(2)
-##' dat <- gamSim(1, n = 400, dist = "normal", scale = 2)
+##' op <- options(digits = 4, cli.unicode = FALSE)
+##' }
+##' dat <- gamSim(1, n = 500, dist = "normal", scale = 2)
 ##' mod <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")
-##' ##'
+##'
 ##' ## point-wise interval
 ##' ci <- confint(mod, parm = "x1", type = "confidence")
-##' head(ci)
-##' ##'
+##' ci
+##'
 ##' ## simultaneous interval for smooth term of x1
+##' \dontshow{
 ##' set.seed(42)
+##' }
 ##' si <- confint(mod, parm = "x1", type = "simultaneous", nsim = 100)
-##' head(si)
+##' si
+##' \dontshow{
+##' options(op)
+##' }
 `confint.gam` <- function(object, parm, level = 0.95, newdata = NULL, n = 200,
                           type = c("confidence", "simultaneous"), nsim = 10000,
                           shift = FALSE, transform = FALSE, unconditional = FALSE,