+++ R/posterior-samples.R
@@ -48,6 +48,7 @@
                                     scale = c("response","linear_predictor"),
                                     freq = FALSE, unconditional = FALSE,
                                     weights = NULL, ncores = 1L, ...) {
+    .NotYetImplemented()
 }
 
 ##' Draw fitted values from the posterior distribution
@@ -88,11 +89,15 @@
 ##'
 ##' @examples
 ##' load_mgcv()
-##' \dontshow{set.seed(2)}
-##' dat <- gamSim(1, n = 400, dist = "normal", scale = 2)
+##' \dontshow{
+##' set.seed(2)
+##' op <- options(cli.unicode = FALSE)
+##' }
+##' dat <- gamSim(1, n = 1000, dist = "normal", scale = 2)
 ##' m1 <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")
 ##'
 ##' fitted_samples(m1, n = 5, seed = 42)
+##' \dontshow{options(op)}
 `fitted_samples.gam` <- function(model, n = 1, newdata, seed,
                                  scale = c("response","linear_predictor"),
                                  freq = FALSE, unconditional = FALSE,
@@ -158,11 +163,15 @@
 ##'
 ##' @examples
 ##' load_mgcv()
-##' \dontshow{set.seed(2)}
-##' dat <- gamSim(1, n = 400, dist = "normal", scale = 2)
+##' \dontshow{
+##' set.seed(2)
+##' op <- options(cli.unicode = FALSE)
+##' }
+##' dat <- gamSim(1, n = 1000, dist = "normal", scale = 2)
 ##' m1 <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")
 ##'
 ##' predicted_samples(m1, n = 5, seed = 42)
+##' \dontshow{options(op)}
 `predicted_samples` <- function(model, ...) {
     UseMethod("predicted_samples")
 }
@@ -240,8 +249,11 @@
 ##'
 ##' @examples
 ##' load_mgcv()
-##' \dontshow{set.seed(2)}
-##' dat <- gamSim(1, n = 400, dist = "normal", scale = 2)
+##' \dontshow{
+##' set.seed(2)
+##' op <- options(cli.unicode = FALSE)
+##' }
+##' dat <- gamSim(1, n = 1000, dist = "normal", scale = 2)
 ##' m1 <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, method = "REML")
 ##'
 ##' smooth_samples(m1, term = "s(x0)", n = 5, seed = 42)
@@ -253,6 +265,7 @@
 ##' ## fit model...
 ##' m2 <- gam(y ~ fac + s(x2, by = fac) + s(x0), data = dat)
 ##' smooth_samples(m2, n = 5, seed = 42)
+##' \dontshow{options(op)}
 `smooth_samples` <- function(model, ...) {
     UseMethod("smooth_samples")
 }